Bugfix. Error while sorting, datetime obj being compared to string.
Refs: #8955

The problem was due to caching and objects in lists being passed by reference.

The lists received from the methods to fetch local and remote news items were
modified after they were received.

The methods themselves were cached.

The result was that on next request, the modified lists are returned from
cache.

The modification being that the 'Date' attr was changed from a string to a
datetime object.